Population Distr. Gender Project for Ireland
This research aims to understand the effects of demographic variables at the regional level by examining the annual distribution of the male and female population in different regions of Ireland. In the study carried out using CRISP-DM methodology, regional population distribution was estimated with machine learning models such as linear regression, Random Forest and Decision Tree, and the results were evaluated with a detailed analysis. Among the machine learning models used in the research, the model with the best performance was Random Forest Regressor. This model achieved lower mean square error (MSE) and higher coefficient of determination (R^2) than other regression models. The findings provide an important basis for understanding the complexity of demographic variables at the regional level and contributing to strategic planning.